requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
google-generativeai>=0.5.2
